--- site/gatsby-config.js
@@ -1,37 +1,0 @@
-module.exports = {
-  siteMetadata: {
-    title: `Capsize`,
-    description: `Flipping how we define typography in CSS.`,
-    author: `@michaeltaranto`,
-  },
-  pathPrefix: `/capsize`,
-  plugins: [
-    `gatsby-plugin-pnpm`,
-    `gatsby-plugin-react-helmet`,
-    {
-      resolve: `gatsby-source-filesystem`,
-      options: {
-        name: `images`,
-        path: `${__dirname}/src/images`,
-      },
-    },
-    `gatsby-transformer-sharp`,
-    `gatsby-plugin-sharp`,
-    {
-      resolve: `gatsby-plugin-manifest`,
-      options: {
-        name: `gatsby-starter-default`,
-        short_name: `capsize`,
-        description: `Flipping how we define typography in CSS.`,
-        start_url: `/`,
-        background_color: `#fff`,
-        theme_color: `#fff`,
-        display: `minimal-ui`,
-        icon: `src/images/capsize-logo.png`, // This path is relative to the root of the site.
-      },
-    },
-    // this (optional) plugin enables Progressive Web App + Offline functionality
-    // To learn more, visit: https://gatsby.dev/offline
-    // `gatsby-plugin-offline`,
-  ],
-};
