+++ site/src/pages/_error/404.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import Layout from '../components/Layout';
-import Seo from '../components/Seo';
+import Layout from '../../components/Layout';
+import Seo from '../../components/Seo';
 
 const NotFoundPage = () => (
   <Layout>
